Extract CryptoBootstrap from PayloadEncryption
PayloadEncryption's companion object was doing three different things: process-global JCE provider mutation (Security.insertProviderAt/removeProvider), Tink registry mutation (AeadConfig.register etc.), and runtime feature-detection state. The class is meant to be a Tink-keyset wrapper, so mixing those responsibilities under its name violated SRP and made it surprising that touching this class altered global JCE state.

Move all the bootstrap to a top-level internal object CryptoBootstrap. PayloadEncryption keeps a single instance init line that calls CryptoBootstrap.ensureInitialized(), preserving the existing 'first construction triggers bootstrap' behavior so direct instantiation sites (tests, OctiServerConnector lazy init, etc.) stay covered without relying on App.onCreate having run first.

RealCryptoCapabilities now delegates to CryptoBootstrap.gcmSivAvailable. PayloadEncryptionTest references the moved helpers via CryptoBootstrap. No behavior change — same providers, same registration, same Tink path. All 13 unit tests still pass.
